requirements.txt
python-chess==1.999
fastapi==0.115.6
uvicorn==0.30.6
